+++ frontend/app/app.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-// const { isLoading } = useLoadingIndicator()
+// const { isLoading } = useLoadingIndicator() пока не делаем не трогать!
 </script>
 
 <template>
@@ -9,5 +9,14 @@
 		<NuxtLayout>
 			<NuxtPage />
 		</NuxtLayout>
+
+		<!-- <Transition name="loader-fade">
+			<div
+				v-if="isLoading"
+				class="fixed inset-0 z-50 flex items-center justify-center bg-[#0a0c0e]/80 backdrop-blur-sm"
+			>
+				<LoadingSpinner size="lg" />
+			</div>
+		</Transition> -->
 	</UApp>
 </template>